dom/tests/mochitest/cowl/labeled_server_communication/test_labeled_ctx_response_bad_priv.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <p id="display"></p>

  <iframe id="cowlframe"></iframe>

  <pre id="test">

  <script class="testbody" type="application/javascript">
  SimpleTest.waitForExplicitFinish();
  SimpleTest.requestFlakyTimeout("Does not trigger any events");

  var completedTests = 0;
  var passedTests = 0;

  var cowlFrame = document.getElementById('cowlframe');
  cowlFrame.src = "http://example.com/tests/dom/tests/mochitest/cowl/postmessage/labeled_bad_priv.html";
  setTimeout(function() {
    var wCowlFrame = SpecialPowers.wrap(cowlFrame.contentWindow);
    console.log(wCowlFrame);

    SimpleTest.finish();
  }, 300);

  function start_tests(iframe) {
    console.log("Start");
    //var wwIframe = SpecialPowers.Cu.waiveXrays(wIframe);

    ok(false, "Should not be able to load labeled response with too strong priv");

  }

  </script>
</pre>
</body>
</html>
